requirements.txt
pyspark>=3.5.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
sqlalchemy>=2.0.23
pandas>=2.1.2